+++ tests/auth/test_happy_path.py
@@ -8,7 +8,7 @@ def test_register_activate_login(client, email, email_user_service):
                           "password2": "123"},
                     follow_redirects=True)
     assert b'Congratulations, you are now a registered user! Check you email for an activation link' in r.data
-    token = email_user_service.call_args.kwargs['token']
+    token = email_user_service.call_args.kwargs['render_params']['url'].split('/')[-1]
     r = client.get(f"auth/activate/{token}", follow_redirects=True)
     assert b'Login' in r.data
     client.post("auth/login",
@@ -23,10 +23,10 @@ def test_password_reset_login(client, email, email_user_service):
                 data={"email": email,
                       "password": "123",
                       "password2": "123"})
-    token = email_user_service.call_args.kwargs['token']
+    token = email_user_service.call_args.kwargs['render_params']['url'].split('/')[-1]
     client.get(f"auth/activate/{token}")
     client.post(f"auth/forgot_password", data={"email": email})
-    token = email_user_service.call_args.kwargs['token']
+    token = email_user_service.call_args.kwargs['render_params']['url'].split('/')[-1]
     client.post(f"auth/password_reset/{token}",
                 data={"password": "321", "password2": "321"})
     client.post("auth/login",