+++ tests/payments/test_customer_portal.py
@@ -6,7 +6,8 @@ def test_customer_portal_url_generation(client, email, email_user_service):
                 data={"email": email,
                       "password": "123",
                       "password2": "123"})
-    token = email_user_service.call_args.kwargs['token']
+    token = email_user_service.call_args.kwargs['render_params']['url'].split('/')[-1]
+
     client.get(f"auth/activate/{token}")
     client.post("auth/login",
                 data={"email": email,