+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Change log:
+19.05.2021
+- add basic email templates
+- change email generation logic
+
 18.05.2021
 - stripe customer portal redirect to manage billing and subscription
 - added requirements.txt
+
 17.05.2021
 - stripe webhook and user payment statuses